Moving edit comment
